requirements.txt
google-cloud-vision==3.4.2
python-multipart

